specify & rename package
