+++ oellrich_graph/core.py
