+++ setup.py
@@ -5,4 +5,5 @@ setup(
     version='1.0.0',
     description='Python implementation of a generic graph class, derived from the C++ implementation of Prof. Martin Oellrich',
     author='Sascha {Thiede, Kohler}',
+    packages=['oellrich_graph'],
 )
