+++ test/test_edge.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from graph.core import Node, Edge
+from oellrich_graph.core import Node, Edge
 
 
 class TestEdge(TestCase):
