+++ test/test_graph.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from graph.core import Graph, Node, Edge
+from oellrich_graph.core import Graph, Node, Edge
 
 
 class TestGraph(TestCase):
