+++ test/test_node.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from graph.core import Node
+from oellrich_graph.core import Node
 
 
 class TestNode(TestCase):
